Account Service Ready
Account Service is Completed

+++ src/services/account/account.service.ts
@@ -1,13 +1,21 @@
 import { Injectable } from '@nestjs/common';
 import { AccountEntity, AccountTypeEntity } from '../../persistence/entities/';
-import { AccountRepository, AccountTypeRepository } from '../../persistence/repositories/';
+import {
+  AccountRepository,
+  AccountTypeRepository,
+} from '../../persistence/repositories/';
 import { AccountModel } from '../../models/';
 import { throws } from 'assert';
+import { PaginationModel } from '../../models/pagination-model.model';
+import { CustomerEntity } from '../../persistence/entities/customer.entity';
+import { AccountController } from '../../controllers/account/account.controller';
 
 @Injectable()
 export class AccountService {
-  constructor(private readonly accountRepository: AccountRepository,
-    private readonly accountTypeRepository: AccountTypeRepository) {}
+  constructor(
+    private readonly accountRepository: AccountRepository,
+    private readonly accountTypeRepository: AccountTypeRepository,
+  ) {}
 
   /**
    * Crear una cuenta
@@ -23,6 +31,21 @@ export class AccountService {
     return this.accountRepository.register(newAccount);
   }
 
+  findAll(pagination: PaginationModel): AccountEntity[] {
+    return this.accountRepository.findAll(pagination);
+  }
+
+  findOneById(accountId: string): AccountEntity {
+    return this.getAccount(accountId);
+  }
+
+  findByCustomer(
+    pagination: PaginationModel,
+    customerId: string,
+  ): AccountEntity[] {
+    return this.accountRepository.findByCustomer(pagination, customerId);
+  }
+
   /**
    * Obtener el balance de una cuenta
    *
@@ -35,56 +58,18 @@ export class AccountService {
   }
 
   /**
-   * Agregar balance a una cuenta
+   * Obtener el tipo de cuenta de una cuenta
    *
    * @param {string} accountId
-   * @param {number} amount
+   * @return {*}  {AccountTypeEntity}
    * @memberof AccountService
    */
-  addBalance(accountId: string, amount: number): void {
-    const account = this.getAccount(accountId);
-    account.balance += amount;
-
-    this.accountRepository.update(accountId , account);
+  getAccountType(accountId: string): AccountTypeEntity {
+    return this.getAccount(accountId).accountType;
   }
 
-  /**
-   * Remover balance de una cuenta
-   *
-   * @param {string} accountId
-   * @param {number} amount
-   * @memberof AccountService
-   */
-  removeBalance(accountId: string, amount: number, removeAll?: boolean): void {
-
-    if(this.verifyAmountIntoBalance(accountId, amount)) throw new Error('Not enough funds');
-
-    if(removeAll) this.cleanBalance(accountId);
-
-    const account = this.getAccount(accountId);
-    account.balance -= amount;
-
-    this.accountRepository.update(accountId , account);
-  }
-
-  private cleanBalance(accountId: string): number {
-    return this.getAccount(accountId).balance = 0; 
-  }
-
-  /**
-   * Verificar la disponibilidad de un monto a retirar en una cuenta
-   *
-   * @param {string} accountId
-   * @param {number} amount
-   * @return {*}  {boolean}
-   * @memberof AccountService
-   */
-  verifyAmountIntoBalance(accountId: string, amount: number): boolean {
-    if(this.getAccount(accountId).balance < amount) {
-        return true
-    }
-
-    return false;
+  getCustomer(accountId: string): CustomerEntity {
+    return this.getAccount(accountId).customer;
   }
 
   /**
@@ -98,26 +83,28 @@ export class AccountService {
     return this.getAccount(accountId).state;
   }
 
-  /**
-   * Cambiar el estado de una cuenta
-   *
-   * @param {string} accountId
-   * @param {boolean} state
-   * @memberof AccountService
-   */
-  changeState(accountId: string, state: boolean): void {
-    this.getAccount(accountId).state = state;
+  updateAccount(accountId: string, newAccount: AccountModel) {
+    let account = this.getAccount(accountId);
+    account = {
+      ...account,
+      ...newAccount,
+    };
+
+    return this.accountRepository.update(accountId, account);
   }
 
   /**
-   * Obtener el tipo de cuenta de una cuenta
+   * Agregar balance a una cuenta
    *
    * @param {string} accountId
-   * @return {*}  {AccountTypeEntity}
+   * @param {number} amount
    * @memberof AccountService
    */
-  getAccountType(accountId: string): AccountTypeEntity {
-    return this.getAccount(accountId).accountType;
+  addBalance(accountId: string, amount: number): void {
+    const account = this.getAccount(accountId);
+    account.balance += amount;
+
+    this.accountRepository.update(accountId, account);
   }
 
   /**
@@ -135,7 +122,23 @@ export class AccountService {
     const account = this.getAccount(accountId);
     account.accountType = this.accountTypeRepository.findOneById(accountTypeId);
 
+    this.accountRepository.update(accountId, account);
+
     return account.accountType;
+  }
+
+  /**
+   * Cambiar el estado de una cuenta
+   *
+   * @param {string} accountId
+   * @param {boolean} state
+   * @memberof AccountService
+   */
+  changeState(accountId: string, state: boolean): void {
+    const account = this.getAccount(accountId);
+    account.state = state;
+
+    this.accountRepository.update(accountId, account);
   }
 
   /**
@@ -145,9 +148,58 @@ export class AccountService {
    * @memberof AccountService
    */
   deleteAccount(accountId: string, soft?: boolean): void {
-    if(soft) this.accountRepository.delete(accountId, soft);
-    
+    if (this.getAccount(accountId).balance != 0)
+      throw new Error(
+        'Cannot Delete this Account. Please transfer your balance to another account',
+      );
+
+    if (soft) this.accountRepository.delete(accountId, soft);
+
     this.accountRepository.delete(accountId);
+  }
+
+  /**
+   * Remover balance de una cuenta
+   *
+   * @param {string} accountId
+   * @param {number} amount
+   * @memberof AccountService
+   */
+  removeBalance(accountId: string, amount: number, removeAll?: boolean): void {
+    if (this.verifyAmountIntoBalance(accountId, amount))
+      throw new Error('Not enough funds');
+
+    if (removeAll) this.cleanBalance(accountId);
+
+    const account = this.getAccount(accountId);
+    account.balance -= amount;
+
+    this.accountRepository.update(accountId, account);
+  }
+
+  private cleanBalance(accountId: string): number {
+    const account = this.getAccount(accountId);
+    account.balance = 0;
+
+    this.accountRepository.update(accountId, account);
+
+    return account.balance;
+  }
+
+  /**
+   * Verificar la disponibilidad de un monto a retirar en una cuenta
+   *
+   * @param {string} accountId
+   * @param {number} amount
+   * @return {*}  {boolean}
+   * @memberof AccountService
+   */
+  verifyAmountIntoBalance(accountId: string, amount: number): boolean {
+    if (this.getAccount(accountId).balance < amount) {
+      return true;
+    }
+
+    return false;
   }
 
   private getAccount(accountId: string): AccountEntity {
